Add files via upload
The basic setup incorporates the bot shell with a HTML frame and a simple UI made with CSS for basic functionality and design. The conversational data needs to be completely re-designed, the UI needs tweaking and the functionality needs to be divided up into a stack of pages instead of the one page of code it is at present. Then we should consider some API integration, a way to present other content such as images video or audio files.